1) Update cross-plat-dev-utils for gcc 4.4.3 2) Fix parameterfileprocessor.c/.h to build with gcc 4.4.3

+++ cross-plat-dev-utils/get_wordsize.c
@@ -16,7 +16,7 @@ Program to return the wordsize in bits of the host machine
 
 int main(void)
 {
-	printf("%lu\n",sizeof(unsigned long));
+	printf("%lu\n",(unsigned long)sizeof(unsigned long));
 	exit(0);
 }
